pyproject.toml
[tool.poetry]
name = "meetup-api-access-token-manager"
version = "0.1.0"
description = "Manages access tokens for the Meetup API."
authors = ["Jan-Benedikt Jagusch <jan.jagusch@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.22"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
black = {version="==19.10b0", allows-prereleases=true}
ipykernel = "^5.1"
pylint = "^2.3"
pytest = "^4.6"
tox = "^3.12"
python-dotenv = "^0.10.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
